Handle error removing pickle file in cleanup.
In 'cleanup_computation' removing pickle file apparently raises
exception sometimes. It may be because 'cleanup_computation' is
being called after it is already removed.

+++ py2/dispy/dispynode.py
@@ -1249,17 +1249,20 @@ class _DispyNode(object):
             logger.debug('pending jobs for computation "%s"/%s: %s',
                          compute.name, compute.id, compute.pending_jobs)
 
+        if self.computations.pop(compute.id, None) is None:
+            logger.warning('Invalid computation "%s" to cleanup ignored' % compute.id)
+            return
+
         pkl_path = os.path.join(self.dest_path_prefix, '%s_%s' % (compute.id, compute.auth))
         if compute.pending_results == 0:
             try:
                 os.remove(pkl_path)
             except:
-                pass
+                logger.warning('Could not remove "%s"' % pkl_path)
         else:
             fd = open(pkl_path, 'wb')
             pickle.dump(compute, fd)
             fd.close()
-        self.computations.pop(compute.id, None)
         try:
             self.scheduler['auth'].remove(compute.auth)
         except ValueError:
